#53 int32_t set in CoordinatusWindow.

+++ gconsole.h
@@ -24,7 +24,7 @@ public:
 
     GConsole(Ui::MainWindow *&ui);
 
-    double_t getPath_mm(uint8_t axis, size_t steps) { return (req_builder->getPath_mm(axis, steps))  ; }
+    double_t getPath_mm(uint8_t axis, int32_t steps) { return (req_builder->getPath_mm(axis, steps))  ; }
 
 
 private slots:
